Add external token which can be used to trigger a retry without having access to the original RequestStateFlow
